pyproject.toml
[tool.poetry]
name = "RASApp"
version = "0.1.0"
description = "RASApp"
authors = ["Rahul Gurujala"]

[tool.poetry.dependencies]
python = "^3.10.3"
fastapi = "0.91.0"
pydantic = "1.10.4"
PyMySQL = "1.0.2"
SQLAlchemy = "2.0.3"
uvicorn = "0.20.0"
Werkzeug = "2.2.2"
python_dateutil = "2.8.2"

[tool.poetry.dev-dependencies]
pytest = "7.2.1"
factory-boy = "3.2.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
